-rebuild part 3 of mongodb schema -implementation of user check

+++ src/main/java/ch/vindthing/controller/AppController.java
@@ -60,17 +60,16 @@ public class AppController {
     @PreAuthorize("hasRole('USER') or hasRole('MODERATOR') or hasRole('ADMIN')")
     public ResponseEntity<?> addItemToStore(@Valid @RequestHeader (name="Authorization") String token,
                                             @RequestBody() ItemAddRequest itemAddRequest) {
-        User user = jwtUtils.getUserFromJwtToken(token);
-
         // Check if store exists
         Store store = storeRepository.findById(itemAddRequest.getStoreId()).
                 orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,
                         "Item Add: Store ID not found: " + itemAddRequest.getStoreId()));
 
         // Usercheck
-        if(!store.getOwner().getId().equals(user.getId())){
-            return ResponseEntity.badRequest().body("Item Add: Not Store of User!");
+        if(!jwtUtils.checkStorePermission(token, store)){
+            return ResponseEntity.badRequest().body("Item Add: No Permission for this Store!");
         }
+
         Item item = new Item(itemAddRequest.getName(), itemAddRequest.getDescription(), itemAddRequest.getQuantity());
         store.getItems().add(item);
         storeRepository.save(store); // Update Store
@@ -85,10 +84,20 @@ public class AppController {
      */
     @RequestMapping("/item/update")
     @PreAuthorize("hasRole('USER') or hasRole('MODERATOR') or hasRole('ADMIN')")
-    public ResponseEntity<?> updateItem(@Valid @RequestBody() ItemUpdateRequest itemUpdateRequest) {
-        // TODO: Only Items/Stores of User
+    public ResponseEntity<?> updateItem(@Valid @RequestHeader (name="Authorization") String token,
+                                        @RequestBody() ItemUpdateRequest itemUpdateRequest) {
         // Find Store by Item ID
         Query query = new Query(Criteria.where("items._id").is(itemUpdateRequest.getId()));
+        try {
+            Store store = mongoTemplate.findOne(query, Store.class);
+            // Usercheck
+            if(!jwtUtils.checkStorePermission(token, store)){
+                return ResponseEntity.badRequest().body("Item Update: No Permission for this Store!");
+            }
+        }catch (Exception e) {
+            return ResponseEntity.badRequest().body("Item Update Failed for ID: " + itemUpdateRequest.getId()
+                    + " Exception: " + e);
+        }
 
         Update update = new Update();
         if(itemUpdateRequest.getName()!=null && !itemUpdateRequest.getName().equals("")){
@@ -128,18 +137,22 @@ public class AppController {
      */
     @RequestMapping("/item/move")
     @PreAuthorize("hasRole('USER') or hasRole('MODERATOR') or hasRole('ADMIN')")
-    public ResponseEntity<?> moveItem(@Valid @RequestBody() ItemMoveRequest itemMoveRequest) {
-        // TODO: Only Items/Stores of User
+    public ResponseEntity<?> moveItem(@Valid @RequestHeader (name="Authorization") String token,
+                                      @RequestBody() ItemMoveRequest itemMoveRequest) {
+        // Find Store and Item by Item ID
         Query query = new Query(Criteria.where("items._id").is(itemMoveRequest.getId()));
         query.fields().include("items.$");
 
-        // Find Item
+        // Find Item and current Store
         Store store;
         Item item;
         try{
             store = mongoTemplate.findOne(query, Store.class);
+            // Usercheck current Store
+            if(!jwtUtils.checkStorePermission(token, store)){
+                return ResponseEntity.badRequest().body("Item Move: No Permission for this Store!");
+            }
             item = store.getItems().get(0);
-            System.out.println("ITEM ID "+item.getId());
         }catch (Exception e) {
             return ResponseEntity.badRequest().body("Item Move: Item ID not found: " + itemMoveRequest.getId()
                     + " Exception: " + e);
@@ -149,6 +162,11 @@ public class AppController {
         Store newStore = storeRepository.findById(itemMoveRequest.getStoreId()).
                 orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,
                         "Item Move: Store ID not found: " + itemMoveRequest.getStoreId()));
+        // Usercheck new Store
+        if(!jwtUtils.checkStorePermission(token, newStore)){
+            return ResponseEntity.badRequest().body("Item Move: No Permission for the new Store!");
+        }
+
         newStore.getItems().add(item);
         storeRepository.save(newStore); // Update Store
 
@@ -173,23 +191,28 @@ public class AppController {
      */
     @RequestMapping("/item/delete")
     @PreAuthorize("hasRole('USER') or hasRole('MODERATOR') or hasRole('ADMIN')")
-    public ResponseEntity<?> deleteItem(@Valid @RequestBody() ItemUpdateRequest itemUpdateRequest) {
-
-        // TODO: only stores of user
+    public ResponseEntity<?> deleteItem(@Valid @RequestHeader (name="Authorization") String token,
+                                        @RequestBody() ItemUpdateRequest itemUpdateRequest) {
+        // Find Store and Item by Item ID
         Query query = new Query(Criteria.where("items._id").is(itemUpdateRequest.getId()));
         query.fields().include("items.$");
 
-        Item newItem;
+        Store store;
+        Item item;
         try{
-            newItem = mongoTemplate.findOne(query, Store.class).getItems().get(0);
-            System.out.println("ITEM ID "+newItem.getId());
+            store = mongoTemplate.findOne(query, Store.class);
+            // Usercheck
+            if(!jwtUtils.checkStorePermission(token, store)){
+                return ResponseEntity.badRequest().body("Item Delete: No Permission for this Store!");
+            }
+            item = store.getItems().get(0);
         }catch (Exception e) {
             return ResponseEntity.badRequest().body("Item Delete: Item ID not found: " + itemUpdateRequest.getId()
                     + " Exception: " + e);
         }
 
         // Delete Item
-        Update update = new Update().pull("items", newItem);
+        Update update = new Update().pull("items", item);
         mongoTemplate.updateFirst(query, update, Store.class);
         return ResponseEntity.ok(new MessageResponse("Item " + itemUpdateRequest.getId() + " successfully deleted!"));
     }
@@ -218,11 +241,15 @@ public class AppController {
      */
     @RequestMapping("/store/update")
     @PreAuthorize("hasRole('USER') or hasRole('MODERATOR') or hasRole('ADMIN')")
-    public ResponseEntity<?> updateStore(@Valid @RequestBody() StoreUpdateRequest storeUpdateRequest) {
-        // TODO: Get User from token, check if store from user
+    public ResponseEntity<?> updateStore(@Valid @RequestHeader (name="Authorization") String token,
+                                         @RequestBody() StoreUpdateRequest storeUpdateRequest) {
         Store store = storeRepository.findById(storeUpdateRequest.getId()).orElseThrow(() ->
                 new ResponseStatusException(HttpStatus.NOT_FOUND,
                         "Store Update: Store ID not found: " + storeUpdateRequest.getId()));
+        // Usercheck
+        if(!jwtUtils.checkStorePermission(token, store)){
+            return ResponseEntity.badRequest().body("Store Update: No Permission for this Store!");
+        }
         if(storeUpdateRequest.getName()!=null && !storeUpdateRequest.getName().equals("")){
             store.setName(storeUpdateRequest.getName());
         }
@@ -245,41 +272,33 @@ public class AppController {
      */
     @RequestMapping("/store/delete")
     @PreAuthorize("hasRole('USER') or hasRole('MODERATOR') or hasRole('ADMIN')")
-    public ResponseEntity<?> deleteStore(@Valid @RequestBody() StoreUpdateRequest storeUpdateRequest) {
-        // TODO: only stores of user, delete all items in store first
+    public ResponseEntity<?> deleteStore(@Valid @RequestHeader (name="Authorization") String token,
+                                         @RequestBody() StoreUpdateRequest storeUpdateRequest) {
         Store store = storeRepository.findById(storeUpdateRequest.getId())
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,
                         "Store Delete: Store ID not found: " + storeUpdateRequest.getId()));
+        // Usercheck
+        if(!jwtUtils.checkStorePermission(token, store)){
+            return ResponseEntity.badRequest().body("Store Delete: No Permission for this Store!");
+        }
         storeRepository.delete(store);
         return ResponseEntity.ok(new MessageResponse("Store deleted!"));
     }
 
     /**
-     * Get all stores and contained items
+     * Get all stores and contained items of user
      * Should be loaded on establishment of websocket
      * @return Stores with items
      */
     @GetMapping("/store/get-all")
     @PreAuthorize("hasRole('USER') or hasRole('MODERATOR') or hasRole('ADMIN')")
-    public ResponseEntity<?> getAllStores() {
+    public ResponseEntity<?> getAllStores(@RequestHeader (name="Authorization") String token) {
+        User user = jwtUtils.getUserFromJwtToken(token);
+
         // TODO: only stores of user
-        List<Store> store = storeRepository.findAll(); //todo stream filter :(
-/*                .stream()
-                .filter(filst -> !filst.getUsers()
-                        .stream()
-                        .filter(filus -> !filus.getId()));*/
+        List<Store> store = storeRepository.findAll();
+
 
         return ResponseEntity.ok(store);
-    }
-
-    @RequestMapping("/store/get-by-name")
-    @PreAuthorize("hasRole('USER') or hasRole('MODERATOR') or hasRole('ADMIN')")
-    @Deprecated
-    public ResponseEntity<?> getStoreByName(@Valid @RequestBody() StoreAddRequest storeRequest) {
-        // TODO: only stores of user
-        Store store = storeRepository.findByName(storeRequest.getName())
-                .orElseThrow(() -> new UsernameNotFoundException("Store Not Found: " + storeRequest.getName()));
-        return ResponseEntity.ok(new StoreResponse(store.getId(), store.getName(), store.getDescription(),
-                store.getLocation(), store.getCreated(), store.getLastedit()));
     }
 }
